Set plugin import settings to be platform dependent.
As of Unity 5 plugins no longer need to be placed within
platform-specific sub-folders of the special Plugins folder. Platform/s
are now set in the Import Settings within the Inspector. This gives the
freedom to either maintain the platform-specific folder convention, or
place the plugins at another convenient location within the Assets
folder.