Using an explicit cast, if it fails, will throw an InvalidCastException, which will un-wind the stack.
The following lines are therefor impossible to hit if fails which means the return if null does not mean anything.

If cast using as ( equivalent to "ARMarker arMarker = target is ARMarker ? (ARMarker)target : (ARMarker)null;" ) then the check for null is possible.

It makes most sense as this is explicitly an Editor for the ARMarker, to just use the explicit cast and go on.